# Add DefaultFocusAdornerProperty and set for themes