--- a/src/Avalonia.Themes.Fluent/Controls/AdornerLayer.xaml
+++ b/src/Avalonia.Themes.Fluent/Controls/AdornerLayer.xaml
@@ -1,14 +1,11 @@
-﻿<Styles xmlns="https://github.com/avaloniaui"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
-  <Styles.Resources>
-    <Thickness x:Key="SystemControlFocusVisualMargin">0</Thickness>
-    <Thickness x:Key="SystemControlFocusVisualPrimaryThickness">2</Thickness>
-    <Thickness x:Key="SystemControlFocusVisualSecondaryThickness">1</Thickness>
-  </Styles.Resources>
-
-  <!--  HighVisibility FocusAdorner  -->
-  <Style Selector=":is(Control)">
-    <Setter Property="FocusAdorner">
+﻿<ResourceDictionary xmlns="https://github.com/avaloniaui"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+  <Thickness x:Key="SystemControlFocusVisualMargin">0</Thickness>
+  <Thickness x:Key="SystemControlFocusVisualPrimaryThickness">2</Thickness>
+  <Thickness x:Key="SystemControlFocusVisualSecondaryThickness">1</Thickness>
+  
+  <ControlTheme x:Key="{x:Type AdornerLayer}" TargetType="AdornerLayer">
+    <Setter Property="DefaultFocusAdorner">
       <FocusAdornerTemplate>
         <Border BorderThickness="{DynamicResource SystemControlFocusVisualPrimaryThickness}"
                 BorderBrush="{DynamicResource SystemControlFocusVisualPrimaryBrush}"
@@ -18,5 +15,5 @@
         </Border>
       </FocusAdornerTemplate>
     </Setter>
-  </Style>
-</Styles>
+  </ControlTheme>
+</ResourceDictionary>