--- a/src/Avalonia.Themes.Simple/Controls/AdornerLayer.xaml
+++ b/src/Avalonia.Themes.Simple/Controls/AdornerLayer.xaml
@@ -1,0 +1,13 @@
+<ResourceDictionary xmlns="https://github.com/avaloniaui"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+  <ControlTheme x:Key="{x:Type AdornerLayer}" TargetType="AdornerLayer">
+    <Setter Property="DefaultFocusAdorner">
+      <FocusAdornerTemplate>
+        <Rectangle Margin="1"
+                   Stroke="Black"
+                   StrokeDashArray="1,2"
+                   StrokeThickness="1" />
+      </FocusAdornerTemplate>
+    </Setter>
+  </ControlTheme>
+</ResourceDictionary>