… src/Avalonia.Themes.Simple/Controls/FocusAdorner.xaml