## What does the pull request do?

Previously for each possible control in the tree we had applied a style activator (perf issue).
Which also wasn't overridable by the custom themes for specific controls (#10158).

## Breaking changes

None.

## Fixed issues

Fixes #10158

@amwx thanks for the idea with default property.

Tried to run benchmarks but got quite big margin of error. Seems to be slightly faster in general.

this PR:
|                   Method |         Mean |      Error |     StdDev |     Gen0 |    Gen1 |  Allocated |
|------------------------- |-------------:|-----------:|-----------:|---------:|--------:|-----------:|
|           CreateCalendar | 4,616.465 us | 46.4592 us | 41.1849 us | 187.5000 | 93.7500 | 3500.03 KB |
| CreateCalendarWithLoaded | 4,470.575 us | 56.1751 us | 52.5462 us | 187.5000 | 93.7500 | 3525.44 KB |
|            CreateControl |     2.966 us |  0.0591 us |  0.0633 us |   0.1488 |       - |    2.73 KB |
|          CreateDecorator |     3.001 us |  0.0576 us |  0.0708 us |   0.1488 |       - |    2.73 KB |
|       CreateScrollViewer |    59.468 us |  1.0875 us |  1.0173 us |   3.9673 |  0.7324 |   73.65 KB |
|             CreateButton |    21.491 us |  0.2585 us |  0.2418 us |   1.3733 |  0.0610 |   25.78 KB |
|            CreateTextBox |   258.334 us |  4.9697 us |  6.8026 us |  11.2305 |  3.4180 |  207.39 KB |

master:
|                   Method |         Mean |      Error |     StdDev |     Gen0 |    Gen1 |  Allocated |
|------------------------- |-------------:|-----------:|-----------:|---------:|--------:|-----------:|
|           CreateCalendar | 4,826.678 us | 65.8151 us | 61.5635 us | 195.3125 | 93.7500 | 3648.21 KB |
| CreateCalendarWithLoaded | 4,470.789 us | 41.3547 us | 36.6598 us | 195.3125 | 93.7500 | 3673.63 KB |
|            CreateControl |     3.260 us |  0.0221 us |  0.0185 us |   0.1678 |       - |    3.11 KB |
|          CreateDecorator |     3.399 us |  0.0292 us |  0.0273 us |   0.1678 |       - |    3.11 KB |
|       CreateScrollViewer |    62.104 us |  1.1922 us |  1.1152 us |   4.0894 |  0.7935 |   75.55 KB |
|             CreateButton |    22.089 us |  0.4224 us |  0.3951 us |   1.4343 |  0.0610 |   26.41 KB |
|            CreateTextBox |   269.385 us |  4.0283 us |  3.7681 us |  11.2305 |  3.4180 |  214.22 KB |

Also, proof of concept of not having any styles in the default themes (aside from control themes) and using shared template for the content control: https://github.com/AvaloniaUI/Avalonia/tree/define-content-control-templates-in-base

|                   Method |         Mean |      Error |     StdDev |     Gen0 |    Gen1 |  Allocated |
|------------------------- |-------------:|-----------:|-----------:|---------:|--------:|-----------:|
|           CreateCalendar | 4,415.020 us | 26.4576 us | 22.0933 us | 179.6875 | 85.9375 | 3406.95 KB |
| CreateCalendarWithLoaded | 4,453.012 us | 42.2125 us | 39.4856 us | 179.6875 | 85.9375 | 3432.37 KB |
|            CreateControl |     2.908 us |  0.0059 us |  0.0052 us |   0.1450 |       - |     2.7 KB |
|          CreateDecorator |     2.937 us |  0.0061 us |  0.0057 us |   0.1450 |       - |     2.7 KB |
|       CreateScrollViewer |    57.115 us |  0.2958 us |  0.2622 us |   3.9673 |  0.7324 |   73.42 KB |
|             CreateButton |    21.304 us |  0.1192 us |  0.1056 us |   1.3733 |       - |    25.7 KB |
|            CreateTextBox |   247.457 us |  2.9554 us |  2.7645 us |  10.7422 |  2.9297 |  206.34 KB |